+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Massive docs update.
 - New printable cases, stands, and reactor parts.
+- Ability to manually reload hardware from the settings menu. (#197)
 - Hardware settings are configurable from the UI in a basic fashion (#141).
 - MicroLab 0.5 simulation configuration (#140).
 - Configurable hardware limits for supported temperatures. Recipes that require values outside of the configured range will no longer run. (#146).